requirements.txt

snowflake-connector-python
pandas
matplotlib
seaborn
